Move expenses into api for easier use
and separation of concerns
